+++ src/fhir/models/careteam-practitioner.ts
@@ -32,12 +32,8 @@ export class CareTeamPractitionerModel extends FHIRModel<fhir4.Practitioner> {
     return this.careTeam.role;
   }
 
-  get qualification() {
+  get specialty() {
     return this.careTeam.practitionerQualification;
-  }
-
-  get telecom() {
-    return this.careTeam.careTeamTelecom;
   }
 
   get status() {